Публичный интерфейс
