+++ Controls/_display/interface/IEditableCollectionItem.ts
@@ -1,12 +1,12 @@
 import {Model} from 'Types/entity';
 
-export interface IEditableCollectionItem<S = any, T extends Model<S> = Model<S>> {
+export interface IEditableCollectionItem<T extends Model = Model> {
     readonly '[Controls/_display/IEditableCollectionItem]': boolean;
     contents: T;
+
     isAdd: boolean;
     addPosition?: 'top' | 'bottom';
 
-    setEditing(editing: boolean, editingContents?: T, silent?: boolean): void;
-    getEditingContents(): T;
-    acceptEditing(): void;
+    setEditing(isEditing: boolean, editingContents?: Model<T>, silent?: boolean): void;
+    acceptChanges(): void;
 }
