+++ Controls/_editInPlace/CollectionEditor.ts
@@ -1,19 +1,37 @@
 import {DestroyableMixin, Model} from 'Types/entity';
-import {ICollectionEditor, ICollectionEditorOptions} from '../interfaces/ICollectionEditor';
-import {TKey, TEditableCollectionItem} from '../interfaces/Types';
+import {RecordSet} from 'Types/collection';
+import {ICollectionEditor, ICollectionEditorOptions} from './interface/ICollectionEditor';
+import {TEditableCollectionItem, TKey} from './Types';
 import {mixin} from 'Types/util';
 
 const ERROR_MSG = {
     ADDING_ITEM_KEY_WAS_NOT_SET: 'Duplicating keys in editable collection. Adding item has the same key as item which is already exists in collection. ' +
         'Maybe you forgot to set a key to previous added item (the requirement of unique keys needs for correct collection operation). ' +
         'In this case you you can set key before edit while prepare adding item, in beforeBeginEdit callback or beforeEndEdit callback.',
-
     ADD_ITEM_KEY_DUPLICATED: 'Duplicating keys in editable collection. Adding item has the same key as item which is already exists in collection.',
-    EDITING_ITEM_MISSED_IN_COLLECTION: ''
+    EDITING_ITEM_MISSED_IN_COLLECTION: 'Item wanna be edit does not exist in source collection.',
+    SOURCE_COLLECTION_MUST_BE_RECORDSET: 'Source collection must be instance of type extended of Types/collection:RecordSet.',
+    HAS_NO_EDITING: 'There is no running edit in collection.'
 };
 
 const ADDING_ITEM_EMPTY_KEY = 'ADD_ITEM_EMPTY_KEY';
 
+/**
+ * Контроллера редактирования коллекции.
+ *
+ * @mixes Controls/_editInPlace/interface/ICollectionEditor
+ * @private
+ * @class Controls/_editInPlace/CollectionEditor
+ * @author Родионов Е.А.
+ */
+/*
+ * Collection editor controller.
+ *
+ * @mixes Controls/_editInPlace/interface/ICollectionEditor
+ * @private
+ * @class Controls/_editInPlace/CollectionEditor
+ * @author Rodionov E.
+ */
 export class CollectionEditor extends mixin<DestroyableMixin>(DestroyableMixin) implements ICollectionEditor {
     private _options: ICollectionEditorOptions;
     private _editingKey: TKey;
@@ -27,55 +45,63 @@ export class CollectionEditor extends mixin<DestroyableMixin>(DestroyableMixin) 
         return this._editingKey;
     }
 
+    getEditingItem(): TEditableCollectionItem {
+        if (this._editingKey === undefined) {
+            return undefined;
+        }
+        return this._options.collection.getItemBySourceKey(this._editingKey);
+    }
+
     updateOptions(options: Partial<ICollectionEditorOptions>): void {
         const oldOptions: Partial<ICollectionEditorOptions> = this._options || {};
         this._options = {collection: options.collection || oldOptions.collection};
+        if (!this._options.collection.getCollection()['[Types/_collection/RecordSet]']) {
+            throw Error(ERROR_MSG.SOURCE_COLLECTION_MUST_BE_RECORDSET);
+        }
     }
 
     edit(item: Model): void {
-        const contents = item.clone();
-        this._editingKey = contents.getKey();
+        this._editingKey = item.getKey();
 
         const collectionItem = this._options.collection.getItemBySourceKey(this._editingKey);
         if (!collectionItem) {
             throw Error(ERROR_MSG.EDITING_ITEM_MISSED_IN_COLLECTION);
         }
 
-        collectionItem.contents.acceptChanges();
-        contents.acceptChanges();
+        collectionItem.acceptChanges();
+        item.acceptChanges();
+        collectionItem.setEditing(true, item);
 
-        collectionItem.setEditing(true, contents);
         this._options.collection.setEditing(true);
-        this._options.collection.getCollection().acceptChanges();
+        (this._options.collection.getCollection() as unknown as RecordSet).acceptChanges();
     }
 
     add(item: Model, addPosition: 'top' | 'bottom' = 'bottom'): void {
-        const contents = item.clone();
-
         // Раньше это разруливалось постфиксами в шаблоне, что не есть норма.
         // Источник не всегда возвращает запись с установленным id, может вернуть undefined, а это уже нормально.
-        // Для одной, добавляемой записи мы обязаны поддержать корректную работу редактировани с пустым ключом, но пользователь
-        // обязан при сохранении наделить ее уникальным ключом, иначе мы упадем с читаемой ошибкой.
-        // таким образом мы не запрещаем редактирование добавляемой записи с пустым ключом, но запрещаем добавление в коллекцию.
-        if (contents.getKey() === undefined) {
-            contents.set(contents.getKeyProperty(), ADDING_ITEM_EMPTY_KEY);
+        // Для одной, добавляемой записи мы обязаны поддержать корректную работу редактировани с пустым ключом,
+        // но пользователь обязан при сохранении наделить ее уникальным ключом, иначе мы упадем с читаемой ошибкой.
+        // таким образом мы не запрещаем редактирование добавляемой записи с пустым ключом,
+        // но запрещаем добавление в коллекцию.
+        if (item.getKey() === undefined) {
+            item.set(item.getKeyProperty(), ADDING_ITEM_EMPTY_KEY);
         }
 
-        this._validateAddingItem(contents);
-        this._editingKey = contents.getKey();
+        this._validateAddingItem(item);
+        this._editingKey = item.getKey();
 
-        const collectionItem = this._options.collection.createItem({contents});
-        collectionItem.contents.acceptChanges();
-        contents.acceptChanges();
+        const collectionItem = this._options.collection.createItem({contents: item});
 
-        collectionItem.setEditing(true, contents);
-        this._options.collection.setEditing(true);
+        collectionItem.acceptChanges();
+        item.acceptChanges();
+        collectionItem.setEditing(true, item);
 
         collectionItem.isAdd = true;
         collectionItem.addPosition = addPosition;
         this._options.collection.setAddingItem(collectionItem);
 
-        this._options.collection.getCollection().acceptChanges();
+        this._options.collection.setEditing(true);
+        (this._options.collection.getCollection() as unknown as RecordSet).acceptChanges();
     }
 
     commit(): void {
@@ -83,12 +109,12 @@ export class CollectionEditor extends mixin<DestroyableMixin>(DestroyableMixin) 
             return;
         }
         const collectionItem = this._options.collection.getItemBySourceKey(this._editingKey);
-        collectionItem.acceptEditing();
+        collectionItem.acceptChanges();
 
         this._options.collection.resetAddingItem();
         collectionItem.setEditing(false, null);
         this._options.collection.setEditing(false);
-        this._options.collection.getCollection().acceptChanges();
+        (this._options.collection.getCollection() as unknown as RecordSet).acceptChanges();
 
         this._editingKey = undefined;
     }
@@ -98,11 +124,14 @@ export class CollectionEditor extends mixin<DestroyableMixin>(DestroyableMixin) 
             return;
         }
         const collectionItem = this._options.collection.getItemBySourceKey(this._editingKey);
+        if (!collectionItem) {
+            throw Error(ERROR_MSG.HAS_NO_EDITING);
+        }
         this._options.collection.resetAddingItem();
 
         collectionItem.setEditing(false, null);
         this._options.collection.setEditing(false);
-        this._options.collection.getCollection().acceptChanges();
+        (this._options.collection.getCollection() as unknown as RecordSet).acceptChanges();
 
         this._editingKey = undefined;
     }
@@ -122,7 +151,8 @@ export class CollectionEditor extends mixin<DestroyableMixin>(DestroyableMixin) 
         if (this._editingKey === undefined) {
             next = collection.getFirst();
         } else {
-            next = direction === 'after' ? collection.getNextByKey(this._editingKey) : collection.getPrevByKey(this._editingKey);
+            next = direction === 'after' ? collection.getNextByKey(this._editingKey) :
+                collection.getPrevByKey(this._editingKey);
         }
 
         while (next && !next['[Controls/_display/IEditableCollectionItem]']) {
